add setup pass conditional

+++ config/functions_win.sh
@@ -13,8 +13,10 @@ hostfile=/mnt/c/Windows/System32/drivers/etc/hosts
 #   None
 #######################################
 function setup_pass() {
-	sudo apt-get update
-	sudo apt-get -y install pass
+	if [[ -z "$(dpkg -s pass | grep Status | grep installed)" ]]; then
+		sudo apt-get update
+		sudo apt-get -y install pass
+	fi
 }
 
 #######################################
